requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.1
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
async-timeout==3.0.1
asyncpg==0.29.0
attrs==24.2.0
certifi==2024.7.4
cffi==1.17.0
chardet==3.0.4
charset-normalizer==3.3.2
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
magic-filter==1.0.12
multidict==4.7.6
outcome==1.3.0.post0
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg2==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.23.1
selenium-async==0.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.32
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.9.4
